Added support for tags in ssh_config in the form of '@tag1 @tag2'.
In order to better manage large numbers of hosts in ssh_config it would be beneficial be able to tag them with project names etc. Now in the 'Host' section of ssh_config one can add a line that starts with a "#" followed (without a spacebar, by a list of tags eg. "#@tag1 @tag2". The 'storm list' command now displays apart from [custom_options] also a [tags] section if flag '-t' is supplied. New 'storm list-tag <tag>' command was added, that displays only Hosts tagged with specific tag, or all tags if no tag is
supplied. The test of 'storm list' was modified to test this new functionality.